refactor: добавила зависимости, удалила неиспользуемые файлы, обновила App.jsx
