--- client/src/actions/index.js
@@ -1,4 +1,0 @@
-import * as types from '../constants/actions';
-
-export const doSomething = ({ type: types.DO_SOMETHING });
-export const doSomethingElse = ({ type: types.DO_SOMETHING_ELSE });
