--- client/src/constants/actions.js
@@ -1,2 +1,0 @@
-export const DO_SOMETHING = 'DO_SOMETHING';
-export const DO_SOMETHING_ELSE = 'DO_SOMETHING_ELSE';
