--- client/src/index.jsx
@@ -1,21 +1,0 @@
-import 'core-js/stable';
-import 'regenerator-runtime/runtime';
-import { render } from 'react-dom';
-import { Provider } from 'react-redux';
-import CssBaseline from '@mui/material/CssBaseline';
-import { ThemeProvider, createTheme } from '@mui/material/styles';
-
-import store from './store';
-import App from './App';
-
-const theme = createTheme();
-
-render(
-  <ThemeProvider theme={theme}>
-    <Provider store={store}>
-      <CssBaseline />
-      <App />
-    </Provider>
-  </ThemeProvider>,
-  document.getElementById('root'),
-);
