--- client/src/reducers/auctions.js
@@ -1,9 +1,0 @@
-const auctions = (state, action) => {
-  switch (action.type) {
-    default: {
-      return { ...state };
-    }
-  }
-};
-
-export default auctions;
