--- client/src/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import auctions from './auctions';
-
-export default combineReducers({
-  auctions,
-});
